Resolve most VDSO comments
